# openapi.yaml
openapi: 3.0.3
info:
  title: Canvastack POSMID API
  description: |-
    API Specification for the Canvastack Point of Sale Multi-tenant ID (POSMID) system.
    This API is designed to be backend-agnostic, following a decoupled, API-First approach.
  version: 1.0.0
servers:
  - url: http://localhost:9000/api/v1
    description: Local Development Server
paths:
...
  /tenants/{tenantId}/products:
    get:
      tags: [Products]
      summary: List all products for a tenant
      description: "Requires authentication and 'products.view' permission within the tenant scope."
      operationId: listProducts
      security: [{ bearerAuth: [] }]
      parameters: [ { $ref: '#/components/parameters/TenantId' } ]
      responses:
        '200':
          description: A list of products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Product' }
    post:
      tags: [Products]
      summary: Create a new product
      description: "Requires authentication and 'products.create' permission within the tenant scope."
      operationId: createProduct
      security: [{ bearerAuth: [] }]
      parameters: [ { $ref: '#/components/parameters/TenantId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductRequest' }
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
  /tenants/{tenantId}/products/{productId}:
    get:
      tags: [Products]
      summary: Get a single product by ID
      operationId: getProductById
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
    put:
      tags: [Products]
      summary: Update an existing product
      operationId: updateProduct
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Product' }
    delete:
      tags: [Products]
      summary: Delete a product
      operationId: deleteProduct
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/ProductId'
      responses:
        '204':
          description: Deleted
  /tenants/{tenantId}/categories:
    get:
      tags: [Products]
      summary: List all categories for a tenant
      description: "Requires authentication and 'products.view' permission within the tenant scope."
      operationId: listCategories
      security: [{ bearerAuth: [] }]
      parameters: [ { $ref: '#/components/parameters/TenantId' } ]
      responses:
        '200':
          description: A list of categories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Category' }
    post:
      tags: [Products]
      summary: Create a new category
      operationId: createCategory
      security: [{ bearerAuth: [] }]
      parameters: [ { $ref: '#/components/parameters/TenantId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryRequest' }
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
  /tenants/{tenantId}/categories/{categoryId}:
    get:
      tags: [Products]
      summary: Get a category by ID
      operationId: getCategoryById
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
    put:
      tags: [Products]
      summary: Update a category
      operationId: updateCategory
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/CategoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CategoryRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Category' }
    delete:
      tags: [Products]
      summary: Delete a category
      operationId: deleteCategory
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/CategoryId'
      responses:
        '204':
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      description: ID of the tenant
      schema: { type: string, format: uuid }
    ProductId:
      name: productId
      in: path
      required: true
      description: ID of the product
      schema: { type: string, format: uuid }
...
    CategoryId:
      name: categoryId
      in: path
      required: true
      description: ID of the category
      schema: { type: string, format: uuid }
...
  schemas:
...
    Product:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        sku: { type: string }
        price: { type: number, format: float }
        stock: { type: integer }
        tenant_id: { type: string, format: uuid }
        category_id: { type: string, format: uuid, nullable: true }
        description: { type: string, nullable: true }
        cost_price: { type: number, format: float, nullable: true }
        created_at: { type: string, example: '2025-01-01 12:00:00' }
        updated_at: { type: string, example: '2025-01-01 12:00:00' }
    ProductRequest:
      type: object
      required: [name, sku, price, stock]
      properties:
        name: { type: string, example: 'Laptop Pro' }
        sku: { type: string, example: 'LP-001' }
        price: { type: number, format: float, example: 1500.99 }
        stock: { type: integer, example: 100 }
        category_id: { type: string, format: uuid, nullable: true }
        description: { type: string, nullable: true }
        cost_price: { type: number, format: float, nullable: true }
...
    Category:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string, nullable: true }
        tenant_id: { type: string, format: uuid }
        created_at: { type: string, example: '2025-01-01 12:00:00' }
        updated_at: { type: string, example: '2025-01-01 12:00:00' }
    CategoryRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, example: 'Elektronik' }
        description: { type: string, nullable: true }
...